Moved utility for child error creation. Added zipper to improve zipWith syntax
