+++ core/src/test/spark_3.0_3.1_3.2/scala/doric/syntax/ArrayColumns3xSpec.scala
@@ -84,10 +84,7 @@ class ArrayColumns3xSpec
         (
             c1,
             c2
-        ) =>
-          colArrayString(c1).zipWith[String, String](concat(_, _))(
-            colArrayString(c2)
-          ),
+        ) => colArrayString(c1).zipWith(colArrayString(c2))(concat(_, _)),
         (c1, c2) => f.zip_with(f.col(c1), f.col(c2), f.concat(_, _)),
         List(Some(Array("ab", "ba", "ce", null)), None, None, None)
       )
